Pokemon Stat Balance Overhaul
Digimon Now have stats based in the realm of pokemon lets go. This was done by having the community picking a pokemon to fit that digimon that also went well with their Digimon Cyber Sleuth Stats if they had them. The game should not play a ton smoother. Also Phoenixmonx's number was in the wrong place that has been fixed in here.